code refactoring [ci skip]

+++ src/shinestacker/project/main_window.py
@@ -78,7 +78,6 @@ class MainWindow(ProjectIOHandler, QMainWindow):
             self.menu_manager.set_enabled_sub_actions_gui)
 
         signal_map = [
-            ('widget_cloned_signal', self.handle_widget_cloned),
             ('widget_pasted_signal', self.handle_widget_pasted),
             ('widget_moved_up_signal', self.handle_widget_moved_up),
             ('widget_moved_down_signal', self.handle_widget_moved_down),
@@ -376,26 +375,13 @@ class MainWindow(ProjectIOHandler, QMainWindow):
 
     def delete_element(self):
         deleted_element, old_selection = self.current_view.delete_element()
-        if deleted_element:
+        if deleted_element and old_selection and old_selection.is_valid():
             for _view_name, view in self.views.items():
                 if view != self.current_view:
                     view.delete_element(selection=old_selection, update_project=False,
                                         confirm=False)
         if self.num_project_jobs() > 0:
             self.menu_manager.delete_element_action.setEnabled(True)
-
-    def handle_widget_cloned(self, indices_tuple):
-        job_idx, action_idx, subaction_idx, widget_type = indices_tuple
-        for _view_name, view in self.views.items():
-            if view != self.sender():
-                state = SelectionState()
-                if widget_type == 'job':
-                    state.set_job(job_idx)
-                elif widget_type == 'action':
-                    state.set_action(job_idx, action_idx)
-                elif widget_type == 'subaction':
-                    state.set_subaction(job_idx, action_idx, subaction_idx)
-                view.clone_element(selection=state, update_project=False, confirm=False)
 
     def handle_widget_pasted(self, indices_tuple):
         job_idx, action_idx, subaction_idx, widget_type = indices_tuple
@@ -488,7 +474,11 @@ class MainWindow(ProjectIOHandler, QMainWindow):
         self.current_view.cut_element()
 
     def clone_element(self):
-        self.current_view.clone_element()
+        success, old_selection = self.current_view.clone_element()
+        if success and old_selection and old_selection.is_valid():
+            for _view_name, view in self.views.items():
+                if view != self.current_view:
+                    view.clone_element(selection=old_selection, update_project=False, confirm=False)
 
     def enable(self):
         self.current_view.enable()
